Pagination + various fixes and improvements (#14)
- In this PR pagination support has been added

![image](https://github.com/user-attachments/assets/613b8e8d-3a0b-41a4-a455-47f286b63dd3)
(you will need to change the Pagination Limit on the RpmLoaderUI or
RpmBasicUi widget to test it)

![image](https://github.com/user-attachments/assets/5d48de66-99f6-4776-a26a-a59991311ed8)
- Big refactoring of how rpm web requests are handled and authenticated
(WebApi, WebApiWithAuth, AssetAPi etc)
- this should fix a number of issues, for example the
delegates/callbacks not running if a request fails
- Added a fallback to cache functionality for Asset Api and
CharacterApi. Instead of manually checking connection, the SDK by
default will try request from the API's and if they fail then it would
fall back to cache if possible.
Also a new request strategy enum. 

![image](https://github.com/user-attachments/assets/b7e33d41-e94f-4721-9ae7-a33d79a8e3ed)
Currently it is used only for AssetApi but it could be expanded upon and
also added to CharacterApi in next iteration